S5PC11X: SOUND: Change audio codec driver to use func ptr for mic bias
In current Herring hw, the gpio used for mic bias is shared between
the audio driver for the main mic bias, and the headset jack driver
for headset mic bias.  This will be split in Rev1.0 HW, but to support
both, change the audio driver to use a function ptr so we can customize
handling based on board rev in the board file.

Also cleaned up problems in error handling in _probe, and get rid of
global pdata.

Change-Id: Ic0b7f656a72cd3e9544bd358e2c949c302912a0b
Signed-off-by: Mike J. Chen <mjchen@sta.samsung.com>