mxt224: TOUCH: Calibrate touchscreen
Touchscreen calibration data. This allows the user to go to the edges of the
screen while maintaining a good linearity and touch response.
This change utilizes the clipping feature of mXT224 for better linearity.
Improved sensitivity and made separate fingers more distinguishable.

'Merge threshold' value is adjusted to prevent the problem separating
a wide object into multiple touches. It is changed from 4 to 14.

Change-Id: I0c8d8ffc65281e50d87a7d1d9d5bd3fa1cc003cd
Signed-off-by: Jinkyu Song <jksong@sta.samsung.com>