ASoC: Replace dead reckoning for WM8994 DC servo with hardware poll
Rather than inserting large sleeps waiting for the DC servo to complete
poll the DC servo status to see if the hardware is done. This ensures that
the delays aren't too long while also avoiding races with the hardware.

In order to aid debug during the transition an info level print displaying
the time taken to complete is left in by default, allowing comparison with
the dead reckinging delays used previously. Overall we save ~5ms in my
tests.

Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>