ASoC: Don't bring up VMID when resuming WM8994
In normal operation the driver leaves VMID disabled when the CODEC is idle
so do this also when coming out of resume. Maintaining VMID consumes a
small amount of power so this should improve battery life, especially when
waking up to perform background syncs with network services.

This will lead to an additional 50ms delay when bringing up the first audio
stream on the device after resume but since we already take a similar hit
for all audio enables after that this should have negligable effect on user
experience.

We should consider also disabling the LDOs as they only take 5ms to bring
up but we need to measure to ensure that power consumption is actually
meaningfully reduce.d

Change-Id: I825e26e5d4916fa7babac93393b795c9ff52d119
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>