S5PC11X: CAMERA: move lens to default position when preview starts
After taking a picture, the focus is maintained.
So if user change from camera mode to video mode after taking
a picture very closely to object,
focus in video mode is not normal.
So move lens to default position when preview starts.
But if latest focus mode is MACRO, the focus should be maintained
as MACRO default after taking a picture.

Change-Id: Ice86ffd9b5932635ba0b4861bc2168559aaafe76
Signed-off-by: kyoungho.yun <kyoungho.yun@samsung.com>