S5PC11X: touchkey: Control the GPJ3[3] to support TFT touchkey LED
Control the GPJ3[3](GPIO_EAR_ADC_SEL) as INPUT/PULLDOWN when the revision is higher than 0.7.
It is not used after 0.7REV and I control it to avoid touchkey LED enabled during the booting time.
If the revision is less than 0.7REV, I control the gpio as OUPUT/HIGH to support the EAR_ADC_SEL.

Signed-off-by: sl47.kim <sl47.kim@samsung.com>
Change-Id: Ic9dbad85d8bd5aa25622bce178827067ca311adb