msm: pm: make temporary mapping used in power collapse large enough
When the system comes up out of power collapse, it relies
on having a temporary mapping where the physical and virtual
addresses are identical, so that the MMU can be turned
on safely.

This mapping is done using 1M pages, but the memory management
code uses the abstraction of 2M pmd segments (so there
are 2 1M pages per pmd).

The code had only mapped in the first of the two 1M pages, and
so if the code for exiting power collapse happened to be in the
second of the two 1M pages, the system would hang
(the PC would be unmapped when the MMU was turned on).

For paranoia's sake, the next 1M page is now also mapped,
in case msm_pm_collapse_exit() is at the very end of the
second 1M page (the actual PC where the MMU is turned on
is 100 bytes or so into the function, but the calculation
of what to map in is done on the entry point).

Change-Id: I17b2014417ad1d3302934ddbbadb0f93c2885575
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>