msm: scm: Fix incorrect cache invalidation
The cache invalidation in scm_call() correctly rounds down the
start address to invalidate the beginning of the cacheline but
doesn't properly round up the 'end' address to make it aligned.
The last chunk of the buffer won't be invalidated when 'end' is
not cacheline size aligned so make sure to invalidate the last
few bytes in such situations.

CRs-Fixed: 281306
Change-Id: I0116f542c288fda65b3d56bb4bed4f5888395e61
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>