mmc: sdhci-s3c: Always update the clock on the first call to set_clock
If the last clock in the array is not the fastest, and the first
call(s) to set_clock selected this clock source, then set_clock
would return with a wrong host->max_clk value and sdhci_set_clock
would program the wrong divider.

The change that initialized cur_clk to the last usable clock
(instead of 0) exposed this bug, but the old code would also have
hit the same problem if the selected (not-fastest) clock was first
in the array.

Change-Id: If1aa5db80af8dcba0199b7c1ed2bb348fa84c9b9
Signed-off-by: Arve Hjønnevåg <arve@android.com>