msm: kgsl: Don't make assumptions about VMA regions
We cannot assume that a VMA region created as a result of an mmap
belongs exclusively to us.  Allow the user to pass the size of
the vmalloc region through the 'gpuaddr' member of the ioctl
struct and do strict checking on the returned vma to make sure
it is valid.

Change-Id: I6504cbc24c291c858403aa518de4ae448dc9f177
Signed-off-by: Wei Zou <wzou@codeaurora.org>
Signed-off-by: Will Tisdale <willtisdale@gmail.com>