msm: msm_hotplug v0.170 This msm hotplug driver is designed to fix some of the shortcomings of the system provided by Qualcomm / HTC for the 8x60 devices. The userspace method by using the mpdecision daemon and a kernel driver which exports runqueue stats is a poor design choice as it is expensive in terms of CPU cycles and it is slow to respond to changes in load, often enabling the second core when the need for it has passed then rapidly offlining it when it could be utilised. The design of this hotplug driver is somewhat different to other ARM MP hotplug implementations. The decision to hotplug is made are by sampling the runqueue at a predetermined interval currently 6ms by calling nr_running(), then a circular buffer is used to average the number of running processes over a predetermined number (currently 5) of sampling periods. A workqueue (msm_hotplug) is utilised to perform the runqueue sampling. If the average number running is over 3.5 processes, then CPU1 is enabled. The sampling rate is then multiplied by 4 to reduce the driver load and to slow down offlining CPU1.
Once the load decreases to 1.5 processes, CPU1 is offlined and the sampling rate returned to default.
The module also hooks into the early suspend handlers, which offline CPU1 as soon as the device
enters early suspend, which in practice means as soon as the display is turned off.

Hotplugging can be disabled on the fly using:
echo 0 > /sys/module/msm_hotplug/parameters/enabled

CPU1 can be forced online on the fly using:
echo 1 > /sys/module/msm_hotplug/parameters/boost

Boost allows CPU1 for be onlined on demand for as long as is required, but the caveat is the device
may not enter power collapse suspend with boost enabled.

Signed-off-by: Will Tisdale <willtisdale@gmail.com>