msm: msm_hotplug v0.172 Increase ENABLE_RUNNING_THRESHOLD to 450. This means that the nr_running must average to 4.5 before the second core is onlined. Short load spikes were causing the CPU to online unnecessarily.
Decrease DISABLE_RUNNING_THRESHOLD to 125.
Due to the sampling rate modifications, we can set this to a lower value
to avoid premature offlining of CPU1.

Increase sampling rate to 10ms.
6ms was unnecessarily low and caused unnecessary hotplug events.
Ideally we want cpufreq to kick in and scale the frequency up to process the
outstanding work before we decide there is too much load. On Android, ondemand
is set to sample at 50ms. msm_hotplug currently uses 6 sample periods to decide
whether to online CPU1, (the previous average and the 5 most recent samples
from the circular buffer) which equates to 60ms with a 10ms sample rate.
This will also be cheaper in terms of CPU cycles.

Reduce sampling rate downsampling multiplier to 2.
As a side effect to increasing the sampling rate, it was taking a long time to
offline CPU1 when it was truly idle. So reduce the multiplier.

Signed-off-by: Will Tisdale <willtisdale@gmail.com>