S5PC11X: CAMERA: delay 150ms before sending macro command in s_parm
The camera driver claims to be in macro focus mode when the lens is
not. When the camera starts preview, default focus mode is auto,
and immediately the focus mode is switched to the new setting.
The second command must be executed with a 150ms delay so that the
camera module can process the first command.

Change-Id: I8e8b5eba8a0cf5e9caa5a370a4495e651bb199b1
Signed-off-by: kyoungho.yun <kyoungho.yun@samsung.com>