cpuidle: idle2: Port of Samsungs LPAUDIO feature to 3.0 kernel The majority of this code is a port from the 2.6.32 P1000 kernel, some of which was required to be rewritten or substantially altered by myself.
Some inspiration was taken from deep idle by ezekeel.

Improvements over deep idle are:
It requires no configuration and is enabled automatically
It does not require the cpu to be at 800MHz for stability

It is not activated itself unless the USB is unplugged, the screen is off and
the PowerManagerService wakelock is held.
  This was a compromise, I wanted an audio wakelock, but as audio uses a partial
  wakelock it isn't visible in the kernel. PowerManagerService deals with the
  partial wakelocks so it is the closest thing to an audio wakelock.

It does not activate itself if USB is plugged in