cpuidle: idle2: v0.121 My reasoning for the likelihood of s5p_idle_bm_check() was flawed.
I hadn't taken into account the fact that if the screen is off and
there are no wakelocks held that the device will enter suspend which
makes cpuidle irrelevant, so therefore I was optimising a less used
code path which made logical sense when I was writing it initially.
So, switch the likely to unlikely in s5p_enter_idle_bm() and the
unlikely to likely in s5p_idle_bm_check(), which will optimise the
code path for idle2 when the screen is turned off. It will save a
couple of cycles.