Fix broken GPU power management
GPU power management control raises the minimum frequency to 200MHz
so that the FSB runs at full speed.

Unfortunately this constantly resets the user policy to 200MHz minimum
which is bad for battery.

This code makes a note of the user policy, and restores it when the
GPU becomes inactive again.